backend/requirements.txt
# Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# LLM & RAG
llama-index-core==0.12.12
openai==1.59.9
httpx==0.28.1

# Vector Database & Retrieval
qdrant-client==1.13.2
fastembed==0.7.4
cohere==5.1.8

# Task Queue
celery==5.4.0
redis==5.2.1

# Database
SQLAlchemy==2.0.19
asyncpg==0.28.0
alembic==1.11.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0

# Google Drive
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1

# Document Parsing
PyMuPDF==1.25.3
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
pytesseract==0.3.13
Pillow==11.1.0

# Utilities
email-validator==2.0.0
python-multipart==0.0.20
python-dotenv==1.0.1
